requirements.txt
requests  
boto3==1.42.81
pandas 
python-dotenv
# dbt-core
# dbt-snowflake
# pytest
